fix: preserve empty run-id selector and make ecc config strict
An explicit --run-id "" collapsed to None inside resolve_run_dir, so every
generated command dropped the selector and fell through to [flow] run:
ecc status --run-id "" inspected runs/default while its bare start_cmd
targeted the configured run, and the run_exists overwrite hint could point
at a different directory than the one just refused. resolve_run_dir now
returns (runs/default, "") for the explicit empty selector and
disclosure_cmd appends --run-id on presence, rendering it as --run-id '';
the displayed name and directory stay default, and absent-selector records
are byte-identical.

ecc config likewise consulted only the ctx-propagated error, which an
explicit selector suppresses, so the step-scoped view exited zero on a
broken [flow] run. The handler now applies the config-owned rule directly
before choosing a view and fails with the reason-bearing config_error
regardless of the selector; the project view also reports invalid_config
instead of raising on an unreadable ecc.toml. The overwrite-guard and
alias-refusal tests move to their own module to stay under the 700-line
review gate.
