+++ chipcompiler/cli/command_handlers/inspect.py
@@ -12,14 +12,15 @@ from chipcompiler.cli.core.output import (
 )
 from chipcompiler.cli.core.records import error_record
 from chipcompiler.cli.core.types import CommandContext, CommandResult
+from chipcompiler.cli.project.config import InvalidFlowRun, config_run_id
 
 
-def _config_error_result(ctx: CommandContext) -> CommandResult:
+def _config_error_result(ctx: CommandContext, reason: str) -> CommandResult:
     return CommandResult.err(
         [
             error_record(
                 "config_error",
-                reason=ctx.config_error,
+                reason=reason,
                 inspect=disclosure_cmd("ecc check", ctx.project),
             )
         ]
@@ -28,7 +29,7 @@ def _config_error_result(ctx: CommandContext) -> CommandResult:
 
 def status(command_input: StatusInput, ctx: CommandContext) -> CommandResult:
     if ctx.config_error:
-        return _config_error_result(ctx)
+        return _config_error_result(ctx, ctx.config_error)
 
     from chipcompiler.cli.inspection.discovery import (
         CORRUPT_FLOW_JSON,
@@ -94,7 +95,7 @@ def status(command_input: StatusInput, ctx: CommandContext) -> CommandResult:
 
 def log(command_input: LogInput, ctx: CommandContext) -> CommandResult:
     if ctx.config_error:
-        return _config_error_result(ctx)
+        return _config_error_result(ctx, ctx.config_error)
 
     from chipcompiler.cli.inspection.discovery import (
         discover_logs,
@@ -218,8 +219,9 @@ def log(command_input: LogInput, ctx: CommandContext) -> CommandResult:
 
 
 def config(command_input: ConfigInput, ctx: CommandContext) -> CommandResult:
-    if ctx.config_error:
-        return _config_error_result(ctx)
+    configured = config_run_id(ctx.project_dir)
+    if isinstance(configured, InvalidFlowRun):
+        return _config_error_result(ctx, configured.problem)
 
     from chipcompiler.cli.inspection.config_view import (
         build_project_config_items,
